requirements.txt
boto3~=1.26.90
psycopg2-binary~=2.9.7
cumulus-message-adapter-python~=2.2.0